// fe/src/lib/api/generated/model/livingArrangementValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type LivingArrangementValue = typeof LivingArrangementValue[keyof typeof LivingArrangementValue];


export const LivingArrangementValue = {
  OWN_HOUSE: 'OWN_HOUSE',
  BOARDING_HOUSE: 'BOARDING_HOUSE',
  WITH_RELATIVES: 'WITH_RELATIVES',
} as const;