// fe/src/lib/api/generated/model/pWDStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type PWDStatusValue = typeof PWDStatusValue[keyof typeof PWDStatusValue];


export const PWDStatusValue = {
  PWD: 'PWD',
  NON_PWD: 'NON_PWD',
  NOT_SPECIFIED: 'NOT_SPECIFIED',
} as const;